fixed routes and inputs

+++ src/pages/NotePage/NotePage.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react"
+import { useEffect, useState, useRef } from "react"
 import { Link, useParams } from "react-router-dom"
 
 import styles from "./NotePage.module.css"
@@ -6,12 +6,27 @@ import styles from "./NotePage.module.css"
 const NotePage = () => {
     const { id } = useParams()
     const [currentNote, setCurrentNote] = useState(null)
+    const titleRef = useRef(null)
+    const contentRef = useRef(null)
 
     useEffect(() => {
         const notes = JSON.parse(localStorage.getItem("notes") || "[]")
         const note = notes.find((note) => note.id === id)
+        // eslint-disable-next-line react-hooks/set-state-in-effect
         setCurrentNote(note)
     }, [id])
+
+    useEffect(() => {
+        if (titleRef.current) {
+            titleRef.current.style.height = "auto"
+            titleRef.current.style.height = titleRef.current.scrollHeight + "px"
+        }
+        if (contentRef.current) {
+            contentRef.current.style.height = "auto"
+            contentRef.current.style.height =
+                contentRef.current.scrollHeight + "px"
+        }
+    }, [currentNote?.title, currentNote?.content])
 
     const save = () => {
         if (!currentNote) return
@@ -22,16 +37,17 @@ const NotePage = () => {
         )
 
         if (currentNote.title.trim() === "") {
-            // const isAgree = confirm(
-            //     "Удалить заметку??? (вы оставили пустой заголовок)",
-            // )
+            const isAgree = confirm(
+                "Удалить заметку??? (вы оставили пустой заголовок)",
+            )
 
-            // if (isAgree) {
-            //     alert("Заметка удалена")
-            // }
-
-            const filteredNotes = updatedNotes.filter((note) => note.id !== id)
-            localStorage.setItem("notes", JSON.stringify(filteredNotes))
+            if (isAgree) {
+                alert("Заметка удалена")
+                const filteredNotes = updatedNotes.filter(
+                    (note) => note.id !== id,
+                )
+                localStorage.setItem("notes", JSON.stringify(filteredNotes))
+            }
         } else {
             localStorage.setItem("notes", JSON.stringify(updatedNotes))
             // alert("Сохранено успешно!")
@@ -53,6 +69,7 @@ const NotePage = () => {
                     Назад
                 </Link>
                 <textarea
+                    ref={titleRef}
                     onChange={(e) =>
                         setCurrentNote({
                             ...currentNote,
@@ -65,6 +82,7 @@ const NotePage = () => {
                     className={styles.titleInput}
                 ></textarea>
                 <textarea
+                    ref={contentRef}
                     onChange={(e) =>
                         setCurrentNote({
                             ...currentNote,
